Migrate iOS example app to RCTReactNativeFactory
