--- example/ios/ReactViewController.swift
@@ -1,32 +1,0 @@
-import Foundation
-import UIKit
-import React
-import React_RCTAppDelegate
-import ReactAppDependencyProvider
-
-class ReactViewController: UIViewController {
-  var reactNativeFactory: RCTReactNativeFactory?
-  var reactNativeFactoryDelegate: RCTReactNativeFactoryDelegate?
-  
-  override func viewDidLoad() {
-    super.viewDidLoad()
-    reactNativeFactoryDelegate = ReactNativeDelegate()
-    reactNativeFactoryDelegate!.dependencyProvider = RCTAppDependencyProvider()
-    reactNativeFactory = RCTReactNativeFactory(delegate: reactNativeFactoryDelegate!)
-    view = reactNativeFactory!.rootViewFactory.view(withModuleName: "TelematicsSdkExample")
-  }
-}
-
-class ReactNativeDelegate: RCTDefaultReactNativeFactoryDelegate {
-  
-  override func sourceURL(for bridge: RCTBridge) -> URL? { self.bundleURL() }
-  
-  override func bundleURL() -> URL? {
-    #if DEBUG
-    RCTBundleURLProvider.sharedSettings().jsBundleURL(forBundleRoot: "index")
-    #else
-    Bundle.main.url(forResource: "main", withExtension: "jsbundle")
-    #endif
-  }
-  
-}
